+++ bin/reinstall-common
@@ -1,0 +1,23 @@
+#!/usr/bin/env bash
+
+# read .env
+set -o allexport
+source .env
+set +o allexport
+
+# rebuild the common package
+cd packages/common && yarn build
+cd ../../
+
+# backend
+yarn remove @ttahub/common
+yarn add file:./packages/common
+# docker-compose run --rm backend yarn remove @ttahub/common
+# docker-compose run --rm backend yarn add file:./packages/common
+
+# frontend
+cd frontend
+yarn remove @ttahub/common
+yarn add file:./../packages/common
+# docker-compose run --rm frontend yarn remove @ttahub/common
+# docker-compose run --rm frontend yarn add file:./packages/common
