+++ docker-compose.override.yml
@@ -39,6 +39,7 @@ services:
       - "3000:3000"
     volumes:
       - "./frontend:/app:rw"
+      - "./packages:/app/packages:ro"
       - "./scripts:/app/scripts"
     environment:
       - BACKEND_PROXY=http://backend:8080
