+++ frontend/src/Constants.js
@@ -103,43 +103,6 @@ export const MY_REPORT_ROLES = [
   'Approver',
 ];
 
-// Note that if this topic list is changed, it needs also to be changed in
-// - src/constants.js
-export const TARGET_POPULATIONS = [
-  'Infants and Toddlers (ages birth to 3)',
-  'Preschool (ages 3-5)',
-  'Pregnant Women',
-  '--------------------',
-  'Affected by Child Welfare Involvement',
-  'Affected by Disaster',
-  'Affected by Substance Use',
-  'Children Experiencing Homelessness',
-  'Children with Disabilities',
-  'Children with Special Health Care Needs',
-  'Dual-Language Learners',
-];
-
-export const ROLES = [
-  'Central Office',
-  'Other Federal Staff',
-  'National Center',
-  'Regional Program Manager',
-  'COR',
-  'Supervisory Program Specialist',
-  'Program Specialist',
-  'Grants Specialist',
-  'Customer Service Contract',
-  'TTAC',
-  'Admin. Assistant',
-  'Early Childhood Manager',
-  'Early Childhood Specialist',
-  'Family Engagement Specialist',
-  'Grantee Specialist Manager',
-  'Grantee Specialist',
-  'Health Specialist',
-  'System Specialist',
-];
-
 export const OTHER_ENTITY_TYPES = [
   'CCDF / Child Care Administrator',
   'Head Start Collaboration Office',
@@ -152,51 +115,6 @@ export const OTHER_ENTITY_TYPES = [
   'State Head Start Association',
   'State Health System',
   'State Professional Development / Continuing Education',
-];
-
-// Note that if this topic list is changed, it needs also to be changed in
-// - src/constants.js
-export const TOPICS = [
-  'Behavioral / Mental Health / Trauma',
-  'Child Screening and Assessment',
-  'CLASS: Classroom Organization',
-  'CLASS: Emotional Support',
-  'CLASS: Instructional Support',
-  'Coaching',
-  'Communication',
-  'Community and Self-Assessment',
-  'Culture & Language',
-  'Curriculum (Instructional or Parenting)',
-  'Data and Evaluation',
-  'Disabilities Services',
-  'ERSEA',
-  'Environmental Health and Safety / EPRR',
-  'Equity',
-  'Facilities',
-  'Family Support Services',
-  'Fiscal / Budget',
-  'Five-Year Grant',
-  'Home Visiting',
-  'Human Resources',
-  'Leadership / Governance',
-  'Learning Environments',
-  'Nutrition',
-  'Ongoing Monitoring Management System',
-  'Oral Health',
-  'Parent and Family Engagement',
-  'Partnerships and Community Engagement',
-  'Physical Health and Screenings',
-  'Pregnancy Services / Expectant Families',
-  'Program Planning and Services',
-  'Quality Improvement Plan / QIP',
-  'Recordkeeping and Reporting',
-  'Safety Practices',
-  'Staff Wellness',
-  'Teaching / Caregiving Practices',
-  'Technology and Information Systems',
-  'Training and Professional Development',
-  'Transition Practices',
-  'Transportation',
 ];
 
 export const REGIONS = [
@@ -254,55 +172,3 @@ const LOCAL_STORAGE_CACHE_NUMBER = '0.2';
 export const LOCAL_STORAGE_DATA_KEY = (id) => `ar-form-data-${id}-${LOCAL_STORAGE_CACHE_NUMBER}`;
 export const LOCAL_STORAGE_ADDITIONAL_DATA_KEY = (id) => `ar-additional-data-${id}-${LOCAL_STORAGE_CACHE_NUMBER}`;
 export const LOCAL_STORAGE_EDITABLE_KEY = (id) => `ar-can-edit-${id}-${LOCAL_STORAGE_CACHE_NUMBER}`;
-
-export const GOAL_CLOSE_REASONS = [
-  'Duplicate goal',
-  'Recipient request',
-  'TTA complete',
-];
-
-export const GOAL_SUSPEND_REASONS = [
-  'Key staff turnover / vacancies',
-  'Recipient request',
-  'Recipient is not responding',
-  'Regional Office request',
-];
-
-/*
-  Please keep in sync with:
-  frontend > src > pages > ActivityReport > constants.js
-  src > constants.js
-*/
-export const RECIPIENT_PARTICIPANTS = [
-  'CEO / CFO / Executive',
-  'Center Director / Site Director',
-  'Coach',
-  'Direct Service: Other',
-  'Family Service Worker / Case Manager',
-  'Fiscal Manager/Team',
-  'Governing Body / Tribal Council / Policy Council',
-  'Home Visitor',
-  'Manager / Coordinator / Specialist',
-  'Parent / Guardian',
-  'Program Director (HS / EHS)',
-  'Program Support / Administrative Assistant',
-  'Teacher / Infant-Toddler Caregiver',
-  'Volunteer',
-];
-
-export const OTHER_ENTITY_PARTICIPANTS = [
-  'HSCO',
-  'Local/State Agency(ies)',
-  'OCC Regional Office',
-  'OHS Regional Office',
-  'Regional Head Start Association',
-  'Regional TTA Team / Specialists',
-  'State Early Learning System',
-  'State Head Start Association',
-  'Other',
-];
-
-export const ALL_PARTICIPANTS = [
-  ...RECIPIENT_PARTICIPANTS,
-  ...OTHER_ENTITY_PARTICIPANTS,
-];
