+++ packages/common/README.md
@@ -1,0 +1,2 @@
+The purpose of this package is to reduce code duplication between the frontend and backend projects.
+
