+++ packages/common/lib/index.js
@@ -1,0 +1,143 @@
+export const RECIPIENT_PARTICIPANTS = [
+    'CEO / CFO / Executive',
+    'Center Director / Site Director',
+    'Coach',
+    'Direct Service: Other',
+    'Family Service Worker / Case Manager',
+    'Fiscal Manager/Team',
+    'Governing Body / Tribal Council / Policy Council',
+    'Home Visitor',
+    'Manager / Coordinator / Specialist',
+    'Parent / Guardian',
+    'Program Director (HS / EHS)',
+    'Program Support / Administrative Assistant',
+    'Teacher / Infant-Toddler Caregiver',
+    'Volunteer',
+];
+export const OTHER_ENTITY_RECIPIENTS = [
+    'HSCO',
+    'Local/State Agency(ies)',
+    'OCC Regional Office',
+    'OHS Regional Office',
+    'Regional Head Start Association',
+    'Regional TTA Team / Specialists',
+    'State Early Learning System',
+    'State Head Start Association',
+    'Other',
+];
+export const ALL_PARTICIPANTS = [
+    ...RECIPIENT_PARTICIPANTS,
+    ...OTHER_ENTITY_RECIPIENTS,
+];
+export const TOPICS = [
+    'Behavioral / Mental Health / Trauma',
+    'Child Screening and Assessment',
+    'CLASS: Classroom Organization',
+    'CLASS: Emotional Support',
+    'CLASS: Instructional Support',
+    'Coaching',
+    'Communication',
+    'Community and Self-Assessment',
+    'Culture & Language',
+    'Curriculum (Instructional or Parenting)',
+    'Data and Evaluation',
+    'Disabilities Services',
+    'ERSEA',
+    'Environmental Health and Safety / EPRR',
+    'Equity',
+    'Facilities',
+    'Family Support Services',
+    'Fiscal / Budget',
+    'Five-Year Grant',
+    'Home Visiting',
+    'Human Resources',
+    'Leadership / Governance',
+    'Learning Environments',
+    'Nutrition',
+    'Ongoing Monitoring Management System',
+    'Oral Health',
+    'Parent and Family Engagement',
+    'Partnerships and Community Engagement',
+    'Physical Health and Screenings',
+    'Pregnancy Services / Expectant Families',
+    'Program Planning and Services',
+    'Quality Improvement Plan / QIP',
+    'Recordkeeping and Reporting',
+    'Safety Practices',
+    'Staff Wellness',
+    'Teaching / Caregiving Practices',
+    'Technology and Information Systems',
+    'Training and Professional Development',
+    'Transition Practices',
+    'Transportation',
+];
+export const REASONS = [
+    'Below Competitive Threshold (CLASS)',
+    'Below Quality Threshold (CLASS)',
+    'Change in Scope',
+    'Child Incidents',
+    'Complaint',
+    'COVID-19 response',
+    'Full Enrollment',
+    'New Recipient',
+    'New Director or Management',
+    'New Program Option',
+    'New Staff / Turnover',
+    'Ongoing Quality Improvement',
+    'Planning/Coordination (also TTA Plan Agreement)',
+    'School Readiness Goals',
+    'Monitoring | Area of Concern',
+    'Monitoring | Noncompliance',
+    'Monitoring | Deficiency',
+];
+export const REPORT_STATUSES = {
+    DRAFT: 'draft',
+    DELETED: 'deleted',
+    SUBMITTED: 'submitted',
+    APPROVED: 'approved',
+    NEEDS_ACTION: 'needs_action',
+};
+export const TARGET_POPULATIONS = [
+    'Infants and Toddlers (ages birth to 3)',
+    'Preschool (ages 3-5)',
+    'Pregnant Women',
+    'Affected by Child Welfare Involvement',
+    'Affected by Disaster',
+    'Affected by Substance Use',
+    'Children Experiencing Homelessness',
+    'Children with Disabilities',
+    'Children with Special Health Care Needs',
+    'Dual-Language Learners',
+];
+export const USER_ROLES = [
+    'Central Office',
+    'Other Federal Staff',
+    'National Center',
+    'Regional Program Manager',
+    'COR',
+    'Supervisory Program Specialist',
+    'Program Specialist',
+    'Grants Specialist',
+    'Customer Service Contract',
+    'TTAC',
+    'Admin. Assistant',
+    'Early Childhood Manager',
+    'Early Childhood Specialist',
+    'Family Engagement Specialist',
+    'Grantee Specialist Manager',
+    'Grantee Specialist',
+    'Health Specialist',
+    'System Specialist',
+];
+export const GOAL_CLOSE_REASONS = [
+    'Duplicate goal',
+    'Recipient request',
+    'TTA complete',
+];
+export const GOAL_SUSPEND_REASONS = [
+    'Key staff turnover / vacancies',
+    'Recipient request',
+    'Recipient is not responding',
+    'Regional Office request',
+];
+//# sourceMappingURL=index.js.map
