+++ packages/common/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@ttahub/common",
+  "version": "1.0.0",
+  "description": "The purpose of this package is to reduce code duplication between the frontend and backend projects.",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "build": "tsc --declaration"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/node": "^18.11.17",
+    "typescript": "^4.9.4"
+  }
+}
