+++ packages/common/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    "module": "es2022",
+    "target": "es2021",
+    "moduleResolution": "node",
+    "outDir": "./lib",
+    "sourceMap": true,
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "allowSyntheticDefaultImports": true,
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "baseUrl": ".",
+    "lib": [
+      "es2021",
+      "dom"
+    ]
+  },
+  "exclude": [
+    "node_modules",
+    "lib"
+  ]
+}
